pyproject.toml
[project]
name = "soda-kr"
version = "0.1.0"
description = "Korean translation of SODA dataset"
requires-python = ">=3.10"
license = {text = "CC-BY-4.0"}
dependencies = [
    "datasets>=2.14.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.66.0",
    "tenacity>=8.2.0",
    "huggingface-hub>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
